+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 * Support primitive `adbPath` expansion ([#3](https://github.com/mpotthoff/vscode-android-webview-debug/pull/3)) by [@buschtoens](https://github.com/buschtoens)
+* Switch to the new [vscode-js-debug](https://github.com/microsoft/vscode-js-debug) extension per default
 
 ## 1.1.2 - 2021-06-18
 * Allow the user to select which page to debug in case multiple are available ([#2](https://github.com/mpotthoff/vscode-android-webview-debug/issues/2))